requirements.txt
numpy
scipy
torch
streamlit
plotly
pandas
PyYAML

